Add TODO item about reducing wrapper object.
