+++ src/model/album.rs
@@ -66,6 +66,7 @@ pub struct FullAlbum {
 /// Full Albums wrapped by Vec object
 ///
 /// [Reference](https://developer.spotify.com/documentation/web-api/reference/albums/get-several-albums/)
+// TODO: Reduce this wrapper object to `Vec<FullAlbum>`
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct FullAlbums {
     pub albums: Vec<FullAlbum>,
@@ -74,6 +75,7 @@ pub struct FullAlbums {
 /// Simplified Albums wrapped by Page object
 ///
 /// [Reference](https://developer.spotify.com/web-api/get-list-new-releases/)
+// TODO: Reduce this wrapper object to `Page<SimplifiedAlbum>`
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct PageSimpliedAlbums {
     pub albums: Page<SimplifiedAlbum>,
