+++ src/model/artist.rs
@@ -49,6 +49,7 @@ pub struct FullArtists {
 /// Full Artists vector wrapped by cursor-based-page object
 ///
 /// [Reference](https://developer.spotify.com/documentation/web-api/reference/follow/get-followed/)
+// TODO: Reduce this wrapper object to `CursorBasedPage<FullArtist>`
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub struct CursorPageFullArtists {
     pub artists: CursorBasedPage<FullArtist>,
