+++ tests/test_models.rs
@@ -410,3 +410,21 @@ fn test_full_playlist() {
     );
     assert_eq!(full_playlist.followers.total, 109);
 }
+
+fn test_devices() {
+    let json_str = r#"
+        {
+            "devices" : [ {
+                "id" : "5fbb3ba6aa454b5534c4ba43a8c7e8e45a63ad0e",
+                "is_active" : false,
+                "is_private_session": true,
+                "is_restricted" : false,
+                "name" : "My fridge",
+                "type" : "Computer",
+                "volume_percent" : 100
+            } ]
+        }
+"#;
+    let payload: DevicePayload = serde_json::from_str(&json_str).unwrap();
+    assert_eq!(payload.devices[0]._type, DeviceType::Computer)
+}
